ref: mv add logic inside users-list; add edit mode
